pridavani vice vizitek najednou

+++ app/model/CardModel.php
@@ -20,45 +20,86 @@ class CardModel {
     public function getAll(){
 	return $this->database->table($this->TableName);
     }
-    
-    
+        
     public function getId($id){
 	return $this->database->table($this->TableName)->get($id);
     }
+    
     /** @return blah bla */
     public function search($keywords){
 	return $this->database->table($this->TableName)->where('surname LIKE ?', '%'.$keywords.'%')->fetch();
-    }
+    }   
     
-    public function zapis($keywords){
-	return $this->database->table('search')->insert([
-	    'keywords' => $keywords,
-	   
-	]);
-    }
-    
-    public function test(){
-	return $this->database->table('search');
+    public function uprav($images){
+	
+	try{
+	    foreach($images as $image){
+		if($image->isImage() AND $image->isOk())
+		{
+		    $extension = pathinfo($image->getSanitizedName(), PATHINFO_EXTENSION);
+		    $imageName = $values->surname;
+		    
+		    /** @TODO vyresit kolize , presun obrazku */    
+		    $img = $image->toImage();
+		    $img->resize(700,400, Image::STRETCH|Image::SHRINK_ONLY);
+		    $img->save(WWW_DIR . $this->UploadThumbPath . $imageName.'.'. $extension);
+		    $img->save(WWW_DIR . $this->UploadPath . $imageName . '.' . $extension);   
+		}
+	    }
+	} catch (Exception $ex) {
+	    $this->presenter->addError($ex->getMessage());
+	}	    
+	
+	/** zpracovani datumu */
+	if($values->date == Null){
+	    $date = date('Y-m-d');
+	}else{
+	    $date = $values->date;
+	}
+	
     }
     
     /** @TODO rozdelit zpracovani obrazku a insert do dtb **/
     public function insert($values){
 	
 	try{
-	    $cards = $values->path;
-	    
+	    $cards = $values->img;
 	    foreach($cards as $card){
 		if($card->isImage() AND $card->isOk())
 		{
 		    $extension = pathinfo($card->getSanitizedName(), PATHINFO_EXTENSION);
-		    $cardName = $values->surname;
+		    if(!$values->surname){
+			$cardName = pathinfo($card->getSanitizedName(), PATHINFO_FILENAME);
+		    }else{
+			$cardName = $values->name . '.' .$extension; 
+		    }
 		    
 		    /** @TODO vyresit kolize , presun obrazku */    
 		    $image = $card->toImage();
 		    $image->resize(700,400, Image::STRETCH|Image::SHRINK_ONLY);
-		    $image->save(WWW_DIR . $this->UploadThumbPath . $cardName.'.'. $extension);
+		    //$image->save(WWW_DIR . $this->UploadThumbPath . $cardName.'.'. $extension);
 		    $image->save(WWW_DIR . $this->UploadPath . $cardName . '.' . $extension);   
 		}
+		
+		    /** zpracovani datumu */
+	    if($values->date == Null){
+		$date = date('Y-m-d');
+	    }else{
+		$date = $values->date;
+	    }
+
+	    $query = $this->database->table($this->TableName)->insert([
+		    'name'=> $values->name,
+		    'surname'=> $cardName,
+		    'workplace'=> $values->workplace,
+		    'project'=>$values->project,
+		    'www'=>$values->www,
+		    'date'=>$date,
+		    'note'=>$values->note,
+		    'img'=>$this->UploadPath . $cardName.'.'.$extension,
+		    'thumb_img'=> $this->UploadThumbPath . $cardName.'.'.$extension,
+		]);
+	
 	    }
 	} catch (Exception $ex) {
 	    $form->addError($ex->getMessage());
@@ -70,18 +111,16 @@ class CardModel {
 	    $date = $values->date;
 	}
 	
-	
 	$query = $this->database->table($this->TableName)->insert([
 		'name'=> $values->name,
 		'surname'=> $cardName,
 		'workplace'=> $values->workplace,
 		'project'=>$values->project,
 		'www'=>$values->www,
-		'photo'=>'',
 		'date'=>$date,
 		'note'=>$values->note,
-		'path'=>'images/cards/'.$cardName.'.'.$extension,
-		'thumb_path'=> 'images/cards/thumbs/'.$cardName.'.'.$extension,
+		'img'=>$this->UploadPath . $cardName.'.'.$extension,
+		'thumb_img'=> $this->UploadThumbPath . $cardName.'.'.$extension,
 	    ]);
 	/** return aby slo presmerovani pri odeslani primo na ukladanou vizitku */
 	return $query;
@@ -89,23 +128,20 @@ class CardModel {
     
     public function update($values){	
 	
-	
-	
-	
+	$this->database->table($this->TableName)->update($values);
 	
     }
     
     public function delete($id){
 	
 	$row = $this->getId($id);
-	$this->database->query('DELETE FROM '.$this->TableName.' WHERE id= '.$id);
-	/*if($row){
-	    unlink(WWW_DIR.$this->UploadPath . $row->path);
-	    unlink(WWW_DIR.$this->UploadThumbPath . $row->thumb_path);
-	    $this->database->query('DELETE FROM '.$this->TableName.' WHERE id= '.$id);
-	    
-	}*/
-	
+	//$this->database->query('DELETE FROM '.$this->TableName.' WHERE id= '.$id);
+	if($row){
+	    unlink(WWW_DIR. $row->img);
+	    //unlink(WWW_DIR.$row->thumb_img);
+	    $this->database->query('DELETE FROM '.$this->TableName.' WHERE id= '.$id);	    
+	}
+	/** @TODO if $image unlink $image */
 	
 	
     }
